# Vectorize rand on uniform distribution for increased performance